Add(index.php,new.php,save_post.php,view.php) : Initial Setup
Created the basic files I need for the initial setup of my blog site. These files are empty for now, but are intended for a foundation of my blog site. Here is the plan :
new.php - a form to add posts
save_post.php - logic for saving a post
index.php - this will display all posts
view.php - this will be to view individual posts