Add/Mod(BlogPosting) : User is now to post save and view a blog post
Set up basic blogging system where the index.php is used as a page to view all posts in posts\ folder. The new.php is used for the creation of a blog post and sends a POST request to save.php. The save_post.php Handles that POST request, and saves that post to a text file labels it a timestamp. The view.php shows the full post with a hyperlink to return to index.php.